Reorganise package files
Moved source files into src/
