+++ .vscode/settings.json
@@ -2,18 +2,19 @@
     "editor.formatOnSave": true,
     "python.formatting.provider": "black",
     "cSpell.words": [
-        "katzplotkinpy",
-        "vprint",
         "AFGEN",
-        "VOR",
-        "SORC",
         "DUB",
         "PHICD",
         "PHICSD",
         "PHILD",
         "PHIQD",
-        "VORING",
+        "SORC",
         "UVLM",
+        "VOR",
+        "VORING",
+        "katzplotkin",
+        "katzplotkinpy",
+        "vprint"
     ],
     "python.pythonPath": "C:\\ProgramData\\Anaconda3\\python.exe",
 }
