+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.2]
+
+### Added
+
+- Unit testing for argparse and version
+
 ## [0.0.1] - 2019.04.03
 
 ### Added
